Ajout du contrôleur de modification profil

+++ src/controllers/utilisateurControllers/UtilisateurSuppression.php
@@ -20,6 +20,18 @@ if (!$isAdmin && !$isOwner) die("Vous ne pouvez pas supprimer un autre utilisate
 $conn = new DatabaseManagement();
 $userCRUD = new UtilisateurCRUD($conn);
 
+$user = $userCRUD->readUserById($userId);
+if (!$user) die("Utilisateur n'existe pas.");
+
+// Supprimer l'image de profil.
+$image = $user->getImageUrl();
+
+if ($image) {
+  $url = dirname(__DIR__) . "../../assets/uploads/profils/$image";
+  unlink($url);
+}
+
+// Supprimer l'utilisateur.
 $userCRUD->deleteUser($userId);
 
 echo "Suppression de l'utilisateur OK.";
